refactor get and gets to return nil; remove getv/s
These changes to get and gets will no longer return errors if the keys
do not exist or if there are no keys existing under a specific
directory. The latter changes are to reduce redundancy, as these functions
are replicated by running `{{ (get "/app").Value }}` in a template, so
these functions are not necessary and reduce the total workload.